improve encode solana: - add cleanup - fetch everytime trade changed - fix fetch aggregator fetch 2 times
